Write a test code to create electronic operators C_up,dw in a systems with N sites as an exercise to learn how to treat fermionic sign.
TODO:
now I will compare the construction of a linear Hubbard chain of few sites
in an ED code (ad-hoc in testEDnsites) and in a DMRG one (testEDkron) using
kron product of fermionic operators. So far without truncation.
This will test the recursive contructing of the blocks and super-block Hamiltonian
without QN conservation.