0.0.1 Working version.
This version solves the 1d Hubbard chain using the infinite DMRG algorithm.
Issues:
The convergence with M is rather slow. Probably this can be improved using
finite DMRG algorithm but for the time being I can not complain.

In the actual implementation we keep track of both SYS and ENV, ie
we do not use any symmetry to recover one from the other.
One possible solution would be to keep track of the first and last site
operators:
Cup_1 -- Cup_N
Cdw_1 -- Cdw_N
P_1 -- P_N
so to correctly grow along either directions.
Yet I need to observe that this is just a small memory issue as
solving for RHO_Ld and RHO_dR is extremely fast compared to construction and
diagonalization of H_sb.

TODO:
1. Try to implement the use of reduced operators exploiting the
separability of the Hilbert spaces for UP and DW spins, as in ED.

2. Implmement quantum number conservation.

Both will help increasing M and overall accuracy.