0.1.0 Stable Version. Milestone 2.
This code is a stable version of the iDMRG for interacting electrons on 1d chain.
The algorithm is general enough but the explicit implementation (test_iDMRG) if
specific of the 1d Hubbard model with nearest-neighbor hopping and local interaction.

More generic implementation can be developed using the same scheme, yet the code
lacks a suitable encapsulation of some generic concepts related to the model and
SuperBlock Hamiltonian construction.