0.1.1 Small Update.
Introduced a Renormalization operation in the BLOCK module, which
performs the rotation + truncation of all the operators in the
operator list  using a given matrix rho_U (obtained from
truncating density matrix rho).

TODO:
This matrix rho_U should be saved in a list within the BLOCK structure
to be used later on to retrieve observables.
In alternative we can introduce a GET_OBSERVABLE part which evaluate
interesting quantities on-the-fly for this iteration==length of the chain.

TODO2:
Check code is working with U>0!