1.0.0 Stable code: Milestone 3
This is a stable version of a simple Lattice DMRG code.
It aims to solve spin-Heisenberg and Hubbard models in 1d,
with local interaction and nearest-neighbour hopping.

It handles Quantum Number Tuple (e.g. Sz or [Nup,Ndw]) conservation.
The code has been benchmarked for Hubbard model U>0 and Heisenberg
models with S=1/2 and S=1.
Further development is in order, see Milestone 4.