Realized that, because the rotation matrices O are sparse, a list of them is simply analog of an operator list, so that to store all the rotations O(i) in the block structure it is enough to add an additional component to the block structure as type(list_operator) :: omatrices
and push them accordingly.

So far however there is an obstacle to this step, because
in idmrg_step we are working with esys <- enlarge(sys) which
however does not store the operator as it is not a deep copy.
On exit with overwrite esys to sys with deep copy. This
erase any saved operator O(i< current_length).

In order to fix this issue and save memory I will try to
avoid copying sys-> esys, env-> eenv and work directly on
the actual block structures (sys,env).