Development + update:
> Test codes have been updated and integrated in each module.
They compile with the pre-processing command -D_TEST (one by
one).

> The entire list of modules have been updated with the goal of
introducing for each class a dedicated copy constructor, i.e.
"=" operation creating a deep copy of the target object.

This is presumably what crashes the finite DMRG algorithm and
in general is a non-safe part of the code.
In doing that I also found some possible bugs, i.e.:
* default constructor (not available anymore)  implementatiion in BLOCKS was wrong.
* the handling of the +qn input in matrix_blocks was completely wrong having an
allocatable definition, which I think is not required.

LIST_SECTORS: +len is now redefined as +dim operation: it returns the
dimension of the basis used in the list of sectors definition.

We are now ready to make a full debug of infinite and finite DMRG.