Possible minor bug fixes from Ricardo Lopes. Intel Fortran compiler raise some errors with the current implemntation.
* intrinsic function dim (list_sectors) requires two arguments and thus is not compatible with the
use we want to make of it (i.e. return the basis dimension of a given block of sectors).
We fixed by removing the `intrinsic dim` statement, de facto defining an independent internal dim function.

* Intel does not like to resolve the overloading of intrinsic operation within the module. So shape(XYZ)
can not be used in the module (hopefully this is the only issue, still need to test to test code).
We fixed by replacing every internal shape() statament with its actual meaning [XYX%N, XYZ%M].

* Related to this was the case of optional variables index in the definition of shape and size procedures
in the MATRIX_BLOCKS objects.

This version should compile with Intel. To be further tested.