Update SYSTEM.f90
Line 798 [ call system("clear") ] may not be conformal with different compilers. (This could lead to error #6413: Invalid global name in this context -- main unit and a variable within that unit sharing the same name)

Line 799 [ call execute_command_line("clear") ] was added to improve the code conformity. (Just a new implementation for executing a command synchronously was adopted.)