Intermediate commit. The kron_hm_1d_2bands reproduces the correct energy and HxV without matrix storage for the 1band model.
H^sb = H^L.x.1 + 1.x.H^R + H^LR
is applied without constructing it and exploiting the quantum numbers.

H^L is decomposed and applied to the columns of \psi_lr
H^R is decomposed and applied to the rows of \psi_lr
H^LR is decomposed into a sum of product \sum_a A^a.x.B^a and applied to \psi_lr
Each A^a and B^a are decomposed in blocks from QNs.

The algorithm is pretty involved but:
1. one shot application of the full Hamiltonian (2 spins, 1band, h.c. activated) is tested
2. evaluation of energy in o-o + o-o 4 sites structure is obtained and is identical to the
one obtained by constructing and storing H^sb using it to apply H^sb \psi_lr.

Next steps:
A. test 2 bands
B. clean and import into the DMRG code to check whether the algorithm is robust to matrix rotations/truncation.