Intermediate commit.
Further simplifying the algorithm A.x.B.
Looking to merge spin and fermions channels:
introduced some changes into SITES and BLOCKS to add further informations:
KeyLink= the prefix of the operators used to connect one block to the other, as used
in H2model.
SiteType= the type of site/block = Spin,Fermion
will be used to automatically select the correct internal algorithm which will
replace H2model.
dQ = the quantum number change with respect to the SuperBlock one as driven by the
off-diagonal blocks in A[q,q+dq] and B[q+dq,q].