Almost final commit. Importd all modifications developed so far (ironically excluding the cmplx scheme!)
The DMRG code has been modified and includes now two execution schemes:
sparse_H = T , F
if T then superblock Hamiltonian is evaluated and used to get groundstate
if F then on-the-fly superblock Hamiltonian is decomposed in
quantum number conserving blocks (diagonal and non-diagonal) and applied
to the Arpack to get groundstate.

The new structure includes:
CONNECT: a module with rather general and abstract evaluation of the
connection Hamiltonians H^{L-R} which is used either in the block extension
and in the evaluation of the superblock matrix if sparse_H = T

SUPERBLOCK: a module containing all the superblock related procedures
this is the workhorse now

DONE: tested fermionic (1,2 bands, half-filled, no interaction) and spin (SUN=2,3) case
for 4 sites (no truncation) using either stand alone and DMRG code procedures.

TODO: test the actual driver for both spin and fermions.