New Version. This version contains a huge update on the code. The code structure has been largely revisited to accomodate
sparse and direct MatVector products and exploit quantum number
conservation in the process (see prev commits).
This code will serve as a platform for MPI development and Multi-Band/SU(N)
use.

TEST: The DMRG spin case for SU(N)=2 has been successfuly tested.
DEBUG: found a couple of bugs in the implementation of the H*v direct
product (mostly related to a bad coding from my side).
*Row/ColOffset was mistakenly indexed with right.block index and not
with the correct superblock index jsb
*The Hright part of the product was using  exchanged dimensions
leading to incorrect vector and row indices. Fixed.

TODO:
1. Test the Fermion case
2. Test Spin and Fermion measure