development version.
Introduced a new measure method using QN decomposition of the operator
and the GS vectors: this enable to avoid storing the GS as block matrix
AND speed up the measure.

The actual implementation (following the HxV direct method) returns
the same numbers as the previous method AND the direct method
using O.x*.I_r or I_l.x*.O which requires more computation time.

The actual method evaluates averages of operators either in the L or R blocks.
Tested for local operators.
TO test:
* non-local operators (e.g. spin-spin)
It does not cover non-local operators across the blocks (L-R), yet this should
be easy to evaluate using the same A.x*.B decomposition in HxV direct.

ACTHUNG: still we have problems with L-R symmetry for the Hubbard model.
Maybe also for the Spin case. Checking this. IT is a fucking long journey
for a simple thing. WTFF....