Intermediate commitL: BUG FIXED
Fucking fermion algebra! The error lies in the incorrect
handling of fermion algebra upon increasing system size.

A patch has been added to enlarge_block.
A more elaborated fix is on the way...
Tests are super promising: energy is now correct, density and double
occupations are very stable with errors of the order of 1d-10.
Mirror symmetry is respected.

More info about the bug in the fixing commit.