3.0.0 NEW STABLE CODE: LIBRARY VERSION
The code now compiles to a library libdmrg.a, it can be used in Fortran
by linking against it and use the related main module.

Test, Documentation and Quick Start examples will be introduced soon.

NOTE: The Spin and Hubbard models in 1d drivers will be places in the
DriverDatabase repository.