Extending sb_get_states to MPI.
Possibly solved the issue by introducing a new vector append in LIST_SECTORS.
For a given QN all states are inserted with one call to avoid loop over
linked list. This can be reused also for the serial case indeed.

All nodes write the states for a given quantum number to a local copy of Astates.
All_reduce in place is performed to sync all copies.
All the states for a given left_qn are inserted by all nodes in their local copy.

Checked and working.