4.2.0 Updated code. * implemented a restart strategy. Now limited to Infinite DMRG algorithm.
- At each step the **initial** blocks L and R are saved to file (according to the values
of block_file and save_all_blocks input variables).
NOTE: the blocks are those entering the DMRG step not the renormalized ones.
This ensures reproducibility at the cost of running again the last loop, which is anyhow
needed to measure as for the construction of the SB basis states and GS wave function.
- For iDMRG: if blocks file with .restart extension are found in the working
directory those are read and used to start the loop using L.length and R.length.

TODO: extend to the Finite DMRG algorithm.
This requires all the blocks from 1 to Ldmrg-1 to be saved to file,
be present and be read. It is just annoying and I am tired.