Using SERIAL get_sb_states and (partial) MPI sb_setup for the H_L,R. There is an issue:
the current implementation of AllGather_MPI for sparse matrices
assume that every node has one element at least, i.e. no empty
nodes are allows. Thus, if MpiSize>Nsb the algorithm stops working.
We need to modify the AllGather to enable existence of nodes that
have no element to share.

This should open to the possibility of performing in parallel also
the construction of A and B (currently serial).