4.9.1 Possible bug fix in sb_get_states.
The last loop, constructing sb_sector and sb_states, has a cycle condition
if(Nk(ql)==0)cycle
This should copy the condition if(.not.right%has(right_qn))cycle as Nk(ql) is
not set for that value of the QN qn. However this requires Nk to be init to 0
which was not done, leading to possible random values in Nk which did not
met condition above.
Error was triggered at random (of course) during testing.
Testing it.