4.10.0 New features. Code extended to possible superc and nonsu2 phases. New input varbiale: dmrg_mode=normal,superc,nonsu2
The "normal" is default. The two new modes corresponds
to a different setup of the local Fock space of the single
dot, which reflects in different quantum numbers (Sz, and Ntot).

Quickly tested for Attractive Hubbard model.
Although it is apparently working there seem to be some issues
with measuring superc observables. Testing it (and learning the
secrets of SC in 1d).