PBC_dev
* Integrating measure inside the dmrg_step, though possible, requires
to propagate each site operator to the latest basis. This can be done in case
one does not want to keep track of the the U.matrices. Thinking about it, not
a priority.

* Quite obviously <H> = E_0 (e grazie ar cazzo), but it is a good sanity check.
Getting <K> is instead more tricky, implemented in the HM driver but get some
discrepancies with respect to ED already for Ldmrg=4. Mmmmhhh.